Move out Defaults and make types concrete
